fastapi-beta/requirements.txt
alembic==1.11.1
anyio==3.7.0
bcrypt==3.1.7
cffi==1.15.1
click==8.1.3
cryptography==3.3.2
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
python-dotenv==1.0.0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
python-dotenv==1.0.0

